(*bst-function-call-sequence*): New. (*relaxed-compilation*): New. (bst-compiler-warning, bst-compile-warning): New.
(pop-form):  Even if NEED-VARIABLE is :IF-SIDE-EFFECTS, we need to
make sure we pop single-value forms only!  (This caused a bug when
compiling a custombib-generated style file.)

(pop-form): Return a fourth value that disambiguates between a Lisp
form NIL and the exhausted-stack situation when :when-empty is NIL.
(pop-single-value-form): Handle this.

(":="): Do nothing for constant variables.  Otherwise, count
assignments.

(get-bst-function): Signal an error if a function is referenced that
could not be compiled.

("IF$"): Handle *RELAXED-COMPILATION*.

(compile-bst-function): HANDLER-BIND for condition
BST-COMPILER-WARNING.

(make-some-variables-lexical): Don't try to make constants lexical.
(make-some-variables-constant): New.