(coerce-form): New. (pop-form): Use it here.
(type-union*): New.

(package-as-body): New keyword arg, expected-result-types.
(package-as-form, package-as-procedure): Pass args to package-as-body.

(":="): Catch assignments to non-variables.  Request the type of the
assigned-to variable when popping the new value.

(compile-literal): New.
(bst-compile-literal): Use it here. Pass remaining args to
package-as-form.

(*silent-compilation*): New variable.
(bst-compile-warning): Don't signal a condition if *silent-compilation*.
("IF$"): Bind *silent-compilation* to true during the first pass.

("IF$"): Fix handling of *relaxed-compilation*.

(assign-loop-variables): New.
(bst-compile-literal-as-while-body): Use it here.
(bst-compile-literal-as-while-predicate): New.
(current-stack-as-values-list): New.
("WHILE$"): Handle the case of loops with complicated data flow.

(compile-funcall): Pass the pop-form-args of a function to pop-form
when popping the arguments.